+++ tests/internal/invalid/upgradeable/use_comprehension/2.py
@@ -1,0 +1,2 @@
+operands = list(map(itemgetter(0), tokens))
+unique_operands = set(map(attrgetter("unique_version"), tokens))