+++ tests/internal/valid/upgradeble/use_comprehension/2.py
@@ -1,0 +1,2 @@
+operands = list(map(basic, tokens))
+unique_operands = set(map(more.basic, tokens))